requirements.txt
streamlit==1.36.0
praw==7.7.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
nltk==3.8.1
requests==2.32.3
python-dotenv==1.0.1
